// src/swagger/transform/models/playlist-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API SPEC FOR MOVESONG FRONTEND
 * REST API SPEC FOR MOVESONG FRONTEND
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PlaylistItemDto
 */
export interface PlaylistItemDto {
    /**
     * 
     * @type {string}
     * @memberof PlaylistItemDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistItemDto
     */
    channelTitle?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistItemDto
     */
    thumbnailUrl?: string;
}