// src/swagger/transform/models/transform-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API SPEC FOR MOVESONG FRONTEND
 * REST API SPEC FOR MOVESONG FRONTEND
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface TransformDto
 */
export interface TransformDto {
    /**
     * 
     * @type {number}
     * @memberof TransformDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    originPlatform?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    destinationPlatform?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    playlistName?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    originPlaylistId?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    destinationPlaylistId?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    movesongEmail?: string;
    /**
     * 
     * @type {number}
     * @memberof TransformDto
     */
    itemCount?: number;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    thumbnailUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformDto
     */
    date?: string;
}